requirements.txt
discord.py~=2.1.0
rethinkdb~=2.4.9
